Refactor scoreboard/grid.
